requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
python-dotenv==1.2.1
openpyxl==3.1.5
pandas==2.3.3
aiofiles==25.1.0